requirements.txt
acryl-datahub==0.12.1.5
acryl-datahub-cloud==0.2.14.3
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
avro==1.11.3
avro-gen3==0.7.11
cached-property==1.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
click-spinner==0.1.10
dataclasses-json==0.6.4
Deprecated==1.2.14
docker==7.0.0
expandvars==0.12.0
frozenlist==1.4.1
graphql-core==3.2.3
humanfriendly==10.0
idna==3.6
ijson==3.2.3
jsonref==1.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
marshmallow==3.21.0
mixpanel==4.10.0
multidict==6.0.5
mypy-extensions==1.0.0
packaging==23.2
progressbar2==4.4.1
psutil==5.9.8
pycarlo==0.9.3
pydantic==2.6.3
pydantic_core==2.16.3
python-box==7.1.1
python-dateutil==2.8.2
python-utils==3.8.2
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-file==2.0.0
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
sentry-sdk==1.40.6
sgqlc==16.3
six==1.16.0
tabulate==0.9.0
termcolor==2.4.0
toml==0.10.2
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
